pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "pyrytiles"
version = "0.1.0"
description = "Python reimplementation of porytiles with CP-SAT solver"
authors = [{ name="Pepas" }]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "Pillow",
    "numpy",
    "ortools",
]

[project.urls]
Homepage = "https://github.com/pgsfernandes/pyrytiles"

license = { text = "MIT" }